requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
opencv-python==4.8.1.78
mediapipe==0.10.7
Pillow==10.0.1
numpy==1.24.3
python-socketio==5.9.0
eventlet==0.33.3
requests==2.31.0
python-dotenv==1.0.0
